requirements.txt
annotated-types==0.7.0
APScheduler==3.11.0
asgiref==3.8.1
beautifulsoup4==4.13.4
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
curl_cffi==0.11.4
cycler==0.12.1
DateTime==5.5
Deprecated==1.2.18
Django==5.2.3
django-apscheduler==0.7.0
django-cors-headers==4.7.0
django-ninja==1.4.3
et_xmlfile==2.0.0
finance-datareader==0.9.96
fonttools==4.58.5
frozendict==2.4.6
idna==3.10
kiwisolver==1.4.8
lxml==5.4.0
matplotlib==3.10.3
multipledispatch==1.0.0
multitasking==0.0.11
mysqlclient==2.2.7
narwhals==1.44.0
numpy==2.3.1
OpenDartReader==0.2.3
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pandas-datareader==0.10.0
peewee==3.18.1
pillow==11.3.0
platformdirs==4.3.8
plotly==6.1.2
protobuf==6.31.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pykrx==1.0.51
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
requests-file==2.1.0
setuptools==80.9.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
websockets==15.0.1
wrapt==1.17.2
xlrd==2.0.2
yfinance==0.2.63
zope.interface==7.2
google-auth==2.34.0
google-auth-oauthlib==1.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
